feat: 마블 상세조회, 미션인증 구현

+++ backend/spring/wassu/src/main/java/com/wassu/wassu/dto/marble/MissionVerifyDTO.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 @Getter
 public class MissionVerifyDTO {
 
-    private long longitude;
-    private long latitude;
+    private double longitude;
+    private double latitude;
 
 }
